requirements.txt
streamlit==1.11.0
pandas==1.4.2
requests==2.28.1
Pillow==9.1.1
